% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invSymmMatrix}
\alias{invSymmMatrix}
\title{Inverse of a matrix}
\usage{
invSymmMatrix(M)
}
\arguments{
\item{M}{a symmetric quadratic matrix}
}
\value{
The inverse of the matrix
}
\description{
Inverse of a matrix
}
\details{
To compute the square root of the matrix, \code{inv_sympd} function of Armadillo library is used.
}
\note{
This function is about 2.5x times faster than R function \code{solve}.

Note that no check is done to test if the matrix M is actually symmetric.
}
\examples{
require(rbenchmark)
M <- matrix(rnorm(1000^2,10,5), nrow=1000)
M_symm <- M\%*\%t(M)
benchmark(solve(M_symm),invSymmMatrix(M_symm),order='relative')

}

