% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mulMatrices}
\alias{mulMatrices}
\title{Multiply two matrices}
\usage{
mulMatrices(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix whose size is compatible with the size of A}
}
\value{
The matrix product A*B.
}
\description{
Multiply two matrices
}
\details{
To compute such product, Armadillo library is used.
}
\note{
This function takes about the same speed as R matrix multiplication.
}
\examples{
require(rbenchmark)
A <- matrix(rnorm(1000*500,10,5), nrow=1000, ncol=500)
B <- matrix(rnorm(1000*500,10,5), nrow=500, ncol=1000)
benchmark(A\%*\%B,mulMatrices(A,B),order='relative')

}

