% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tMatrix}
\alias{tMatrix}
\title{Transpose of a matrix}
\usage{
tMatrix(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
Its transpose
}
\description{
Transpose of a matrix
}
\details{
To compute the transpose matrix, Armadillo library is used.
}
\note{
This function is about 4 times slower than R function \code{t}.
}
\examples{
require(rbenchmark)
M <- matrix(rnorm(1000*5000,10,5), nrow=1000)
benchmark(t(M),tMatrix(M),order='relative')

}

